requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
Werkzeug==3.1.3
